[혜원] 종목 별 게시물 리스트 조회 기능 구현

+++ src/main/java/com/monstersinc/stock101/community/controller/CommunityController.java
@@ -31,7 +31,7 @@ public class CommunityController {
     public ResponseEntity<BaseResponseDto<PostResponseDto>> create(
             @Valid @RequestBody PostRequestDto requestDto) {
 
-        int newId = communityService.save(requestDto);
+        long newId = communityService.save(requestDto);
         PostResponseDto body = communityService.getAPost(newId);
 
         return ResponseEntity.status(HttpStatus.CREATED)
